util: expose toUSVString
Expose toUSVString to it can be used by user libraries.

Refs: https://github.com/nodejs/undici/pull/986
