+++ lib/internal/url.js
@@ -19,7 +19,6 @@ const {
   ReflectApply,
   ReflectGetOwnPropertyDescriptor,
   ReflectOwnKeys,
-  RegExpPrototypeExec,
   String,
   StringPrototypeCharCodeAt,
   StringPrototypeIncludes,
@@ -40,7 +39,12 @@ const {
   isHexTable
 } = require('internal/querystring');
 
-const { getConstructorOf, removeColors } = require('internal/util');
+const {
+  getConstructorOf,
+  removeColors,
+  toUSVString
+} = require('internal/util');
+
 const {
   ERR_ARG_NOT_ITERABLE,
   ERR_INVALID_ARG_TYPE,
@@ -79,7 +83,6 @@ const {
   domainToASCII: _domainToASCII,
   domainToUnicode: _domainToUnicode,
   encodeAuth,
-  toUSVString: _toUSVString,
   parse,
   setURLConstructor,
   URL_FLAGS_CANNOT_BE_BASE,
@@ -112,18 +115,6 @@ const kFormat = Symbol('format');
 const IteratorPrototype = ObjectGetPrototypeOf(
   ObjectGetPrototypeOf([][SymbolIterator]())
 );
-
-const unpairedSurrogateRe =
-    /(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])/;
-function toUSVString(val) {
-  const str = `${val}`;
-  // As of V8 5.5, `str.search()` (and `unpairedSurrogateRe[@@search]()`) are
-  // slower than `unpairedSurrogateRe.exec()`.
-  const match = RegExpPrototypeExec(unpairedSurrogateRe, str);
-  if (!match)
-    return str;
-  return _toUSVString(str, match.index);
-}
 
 // Refs: https://html.spec.whatwg.org/multipage/browsers.html#concept-origin-opaque
 const kOpaqueOrigin = 'null';
