+++ lib/util.js
@@ -72,7 +72,8 @@ const {
   deprecate,
   getSystemErrorMap,
   getSystemErrorName: internalErrorName,
-  promisify
+  promisify,
+  toUSVString
 } = require('internal/util');
 
 let internalDeepEqual;
@@ -368,6 +369,7 @@ module.exports = {
   isPrimitive,
   log,
   promisify,
+  toUSVString,
   TextDecoder,
   TextEncoder,
   types
